gem5-SALAM-gui/requirements.txt
PySide6>=6.2.0
PyYAML>=6.0

# Optional dependencies for visualization
pydot>=1.4.0
graphviz>=0.20
